// lib/generated/prisma/enums.ts
/* !!! This is code generated by Prisma. Do not edit directly. !!! */
/* eslint-disable */
// biome-ignore-all lint: generated file
// @ts-nocheck 
/*
* This file exports all enum related types from the schema.
*
* 🟢 You can import this file directly.
*/

export const OrgPlan = {
  FREE: 'FREE',
  TEAM: 'TEAM',
  BUSINESS: 'BUSINESS'
} as const

export type OrgPlan = (typeof OrgPlan)[keyof typeof OrgPlan]


export const OrgRole = {
  OWNER: 'OWNER',
  ADMIN: 'ADMIN',
  MEMBER: 'MEMBER'
} as const

export type OrgRole = (typeof OrgRole)[keyof typeof OrgRole]


export const MeetingStatus = {
  PROCESSING: 'PROCESSING',
  READY: 'READY',
  FAILED: 'FAILED'
} as const

export type MeetingStatus = (typeof MeetingStatus)[keyof typeof MeetingStatus]


export const ActionStatus = {
  OPEN: 'OPEN',
  DONE: 'DONE'
} as const

export type ActionStatus = (typeof ActionStatus)[keyof typeof ActionStatus]